fix: Исправить ошибки компиляции Flutter
- Исправлена синтаксическая ошибка в VenueModel (лишняя кавычка)
- Добавлена null-safety проверка для priceWeekday в CourtDetailScreen
- Исправлено приведение типов для pricePerPlayer в SimpleGameTypeScreen

+++ lib/screens/simple_game_type_screen.dart
@@ -119,8 +119,8 @@ class _SimpleGameTypeScreenState extends State<SimpleGameTypeScreen> {
       // Calculate price per player for open games
       final totalPrice = widget.price;
       final pricePerPlayer = selectedGameType == 'open' 
-          ? totalPrice / selectedPlayersCount 
-          : totalPrice;
+          ? totalPrice.toDouble() / selectedPlayersCount 
+          : totalPrice.toDouble();
       
       // Создаем бронирование
       final bookingId = await BookingService().createBooking(
